+++ packages/imperative/src/config/cmd/convert-profiles/convert-profiles.definition.ts
@@ -25,12 +25,20 @@ export const convertProfilesDefinition: ICommandDefinition = {
     summary: "Convert profiles to team config",
     description: `Convert v1 profiles to a global ${ImperativeConfig.instance.rootCommandName}.config.json file.`,
     options: [{
-        name: "force",
-        description: "Bypass prompt for user confirmation before profile conversion.",
+        name: "prompt",
+        description: "Prompt for confirmation. Use --no-prompt to disable prompting.",
+        type: "boolean",
+        defaultValue: true
+    }, {
+        name: "delete",
+        description: "Delete the existing profiles on disk and any securely stored secrets.",
         type: "boolean"
     }],
     examples: [{
         description: "Convert profiles to team config without prompting",
-        options: "--force"
+        options: "--no-prompt"
+    }, {
+        description: "Convert profiles to team config and delete the old profiles",
+        options: "--delete"
     }]
 };